refactor(classes): extract CLASS_NAME paraglide map to shared helper
Four call sites duplicated the same Record<CharacterClassId, () => string>
inline. Extracts to src/game/classes/i18n.ts → getClassDisplayName(classId).
Fallback unified on m.unknown_class() (StatusCard previously fell back to
literal "Unknown"; admin/users to raw classId — both now translate).

Also drops the now-dead classDef prop on StatusCard (only consumer was the
replaced name resolution) and the matching classDef={classDef} arg in
world.tsx.
